refactor client jest configs to re-use base config
